Finish redisign the project
