+++ encryption/App_Start/RouteConfig.cs
@@ -10,7 +10,7 @@ namespace encryption
             routes.IgnoreRoute("{resource}.axd/{*pathInfo}");
             routes.MapRoute(name: "Default",
                             url: "{controller}/{action}/{id}",
-                            defaults: new { controller = "Encrypt", action = "Index", id = UrlParameter.Optional });
+                            defaults: new { controller = "Transposition", action = "Index", id = UrlParameter.Optional });
         }
     }
 }
