+++ encryption/Views/Shared/Layout.cshtml
@@ -17,8 +17,9 @@
             <input id="bmenub" type="checkbox" class="show">
             <label for="bmenub" class="burger pseudo button">&#9776;</label>
             <div class="menu">
-                <a href="/Encrypt" class="button">Encriptar</a>
-                <a href="/Decrypt" class="button">Desencriptar</a>
+                <a href="/Transposition" class="button">Transposición</a>
+                <a href="/SDES" class="button">S-DES</a>
+                <a href="/RSA" class="button">RSA</a>
             </div>
         </nav>
     </div>
