+++ encryption/Views/Transposition/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@{
-    ViewBag.Title = "Encriptar Archivos";
+    ViewBag.Title = "Cifrado por Transposicion";
     Layout = "~/Views/Shared/Layout.cshtml";
 }
 <input type="hidden" id="state" value="@ViewBag.Message" />
@@ -8,8 +8,8 @@
     <p id="message"></p>
 </div>
 <div style="text-align: center">
-    <form action="/Encrypt/Encrypt" method="post" enctype="multipart/form-data" style="display: inline-block">
-        <h1>Encriptar</h1>
+    <form enctype="multipart/form-data" style="display: inline-block">
+        <h1>Cifrado por Trasnposición</h1>
         <input type="file" class="input-file" name="file" id="file" accept=".txt">
         <label for="file" class="file">
             <i class="fas fa-file-upload"></i><br />Selecciona un archivo
@@ -23,12 +23,11 @@
         <label for="radio2" class="checkable">Cesar</label><br />
         <input id="radio3" type='radio' name="cipher" value="Spiral">
         <label for="radio3" class="checkable">Espiral</label><br />
-        <input id="radio4" type='radio' name="cipher" value="SDES">
-        <label for="radio4" class="checkable">SDES</label><br />
         <div id="keys" style="display: inline-block">
             <p>Clave: </p>
             <input type="number" name="key" min="1" max="21474836" placeholder="Niveles de profundidad" required />
         </div><br /><br />
-        <button type="submit">Encriptar</button>
+        <button type="submit" formaction="/Transposition/Encrypt" formmethod="post">Encriptar</button>
+        <button type="submit" formaction="/Transposition/Decrypt" formmethod="post">Desencriptar</button>
     </form>
 </div>
