+++ encryption/encryption.csproj
@@ -113,8 +113,8 @@
   <ItemGroup>
     <Compile Include="App_Start\BundleConfig.cs" />
     <Compile Include="App_Start\RouteConfig.cs" />
-    <Compile Include="Controllers\DecryptController.cs" />
-    <Compile Include="Controllers\EncryptController.cs" />
+    <Compile Include="Controllers\SDESController.cs" />
+    <Compile Include="Controllers\TranspositionController.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
@@ -130,8 +130,8 @@
     <Content Include="Views\web.config" />
     <None Include="packages.config" />
     <Content Include="Views\Shared\Layout.cshtml" />
-    <Content Include="Views\Encrypt\Index.cshtml" />
-    <Content Include="Views\Decrypt\Index.cshtml" />
+    <Content Include="Views\Transposition\Index.cshtml" />
+    <Content Include="Views\SDES\Index.cshtml" />
     <None Include="Web.Debug.config">
       <DependentUpon>Web.config</DependentUpon>
     </None>
